chess-ai-app/requirements.txt
flask
flask-cors
torch
numpy
python-chess
gunicorn